% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQHI.R
\name{AQHI}
\alias{AQHI}
\title{Calculate the Canadian AQHI from hourly PM2.5, NO2, and O3 observations}
\usage{
AQHI(dates, pm25_1hr_ugm3, no2_1hr_ppb = NA, o3_1hr_ppb = NA, verbose = TRUE)
}
\arguments{
\item{dates}{Vector of hourly datetimes corresponding to observations. Date gaps will be filled automatically.}

\item{pm25_1hr_ugm3}{Numeric vector of hourly mean fine particulate matter (PM2.5) concentrations (ug/m^3).}

\item{no2_1hr_ppb}{(Optional). Numeric vector of hourly mean nitrogen dioxide (NO2) concentrations (ppb). If not provided AQHI+ will be calculated from PM2.5 only.}

\item{o3_1hr_ppb}{(Optional). Numeric vector of hourly mean ozone (O3) concentrations (ppb). If not provided AQHI+ will be calculated from PM2.5 only.}

\item{verbose}{(Optional). A single logical (TRUE/FALSE) value indicating if non-critical warnings/messages should be displayed. Default is TRUE}
}
\value{
A tibble (data.frame) with columns (\emph{if all 3 pollutants provided):
date, pm25, o3}, no2*, pm25_rolling_3hr*, o3_rolling_3hr*, o3_rolling_3hr*,
AQHI, AQHI_plus, risk, high_risk_pop_message, general_pop_message, AQHI_plus_exceeds_AQHI*
and potentially more rows than \code{length(dates)} (due to any missing hours being filled with NA values).
}
\description{
The Canadian Air Quality Health Index (AQHI) combines the health risk of
fine particulate matter (PM2.5), ozone (O3), and nitrogen dioxide (NO2) on a scale from 1-10 (+).
The AQHI is "the sum of excess mortality risk associated with individual pollutants
from a time-series analysis of air pollution and mortality in Canadian cities,
adjusted to a 0–10 scale, and calculated hourly on the basis of trailing 3-hr average pollutant concentrations."

The AQHI is overriden by the Canadian AQHI+ if the AQHI+ exceeds the AQHI for a particular hour.
The AQHI+ is a modification of the Canadian Air Quality Health Index (AQHI).
AQHI+ only uses fine particulate matter (PM2.5) instead of the combination of PM2.5, ozone (O3), and nitrogen dioxide (NO2).
Unlike the AQHI which uses 3-hour mean averages, AQHI+ is calculated using hourly means.

The AQHI was originally published by Steib et. al in 2008,
and has been adopted by all Canadian provinces/territories
(except Quebec where they use the AQI instead of the AQHI/AQHI+).
}
\examples{
obs <- data.frame(
  date = seq(
    lubridate::ymd_h("2024-01-01 00"),
    lubridate::ymd_h("2024-01-01 23"), "1 hours"
  ),
  pm25 = sample(1:150, 24), o3 = sample(1:150, 24), no2 = sample(1:150, 24)
)
AQHI(
  dates = obs$date, pm25_1hr_ugm3 = obs$pm25,
  o3_1hr_ppb = obs$o3, no2_1hr_ppb = obs$no2
)

AQHI(dates = obs$date, pm25_1hr_ugm3 = obs$pm25) # Returns AQHI+
}
\references{
\url{https://doi.org/10.3155/1047-3289.58.3.435}
}
\seealso{
Other Canadian Air Quality: 
\code{\link{AQHI_plus}()}

Other Air Quality Standards: 
\code{\link{AQHI_plus}()}
}
\concept{Air Quality Standards}
\concept{Canadian Air Quality}
