% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{data}
\name{AQHI_colours}
\alias{AQHI_colours}
\title{AQHI_colours}
\format{
A named character vector.
}
\source{
The AQHI colours were derived from \href{https://www.canada.ca/en/environment-climate-change/services/air-quality-health-index/about.html}{Environment and Climate Change Canada}.
}
\usage{
AQHI_colours
}
\description{
A named character vector of hexidecimal codes used to represent the AQHI levels.
}
\keyword{datasets}
