% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{data}
\name{AQHI_health_messages}
\alias{AQHI_health_messages}
\title{AQHI_health_messages}
\format{
A namedad list of tibbles, 1 tibble per language.
}
\source{
\href{https://www.canada.ca/en/environment-climate-change/services/air-quality-health-index/understanding-messages.html}{Environment and Climate Change Canada}.
}
\usage{
AQHI_health_messages
}
\description{
A named list of the health messages for the AQHI risk categories for both English ("en") and French ("fr").
}
\keyword{datasets}
