% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQHI_plus.R
\name{AQHI_plus}
\alias{AQHI_plus}
\title{Calculate the Canadian AQHI+ from hourly \ifelse{html}{\out{PM<sub>2.5</sub>}}{\eqn{PM_{2.5}}} concentrations}
\usage{
AQHI_plus(
  pm25_1hr_ugm3,
  min_allowed_pm25 = 0,
  detailed = TRUE,
  language = "en"
)
}
\arguments{
\item{pm25_1hr_ugm3}{A numeric vector of hourly mean \ifelse{html}{\out{PM<sub>2.5</sub>}}{\eqn{PM_{2.5}}} concentrations (units = \ifelse{html}{\out{&mu;g m<sup>-3</sup>}}{\eqn{ug m^{-3}}}).}

\item{min_allowed_pm25}{(Optional).
A single numeric value indicating the minimum allowed concentration.
All values in \code{pm25_1hr_ugm3} less than this will be replaced with NA.
Default is 0 \ifelse{html}{\out{&mu;g m<sup>-3</sup>}}{\eqn{ug m^{-3}}}.}

\item{detailed}{(Optional).
A single logical value indicating if a tibble with levels, risk categories, health messages, etc should be returned.
If FALSE only the levels will be returned.
Default is TRUE.}

\item{language}{(Optional).
A single character value indicating the language to use for risk levels and health messaging.
Must be either "en" (English) or "fr" (French). Not case sensitive.
Default is "en".}
}
\value{
If \code{detailed} is TRUE:
\itemize{
\item A tibble (data.frame) with columns detailing the AQHI+, risk level, health messages, etc
}

If \code{detailed} is FALSE:
\itemize{
\item A factor vector of AQHI+ levels with \code{length(pm25_1hr_ugm3)} elements
}
}
\description{
The Canadian AQHI+ is a modification of the Canadian Air Quality Health Index (AQHI).
AQHI+ only uses fine particulate matter (\ifelse{html}{\out{PM<sub>2.5</sub>}}{\eqn{PM_{2.5}}}) instead of the combination of
\ifelse{html}{\out{PM<sub>2.5</sub>}}{\eqn{PM_{2.5}}}, ozone (\ifelse{html}{\out{O<sub>3</sub>}}{\eqn{O_3}}), and nitrogen dioxide (\ifelse{html}{\out{NO<sub>2</sub>}}{\eqn{NO_2}}).
Unlike the AQHI which uses 3-hourly mean averages,
AQHI+ is calculated using hourly mean averages.
The AQHI+ overrides the AQHI if it exceeds the AQHI for a particular hour.

AQHI+ categorizes hourly \ifelse{html}{\out{PM<sub>2.5</sub>}}{\eqn{PM_{2.5}}} concentrations into
10 equal bins from 0 to 100 \ifelse{html}{\out{&mu;g m<sup>-3</sup>}}{\eqn{ug m^{-3}}} (AQHI+ 1-10),
and assigns "+" to values greater than 100 \ifelse{html}{\out{&mu;g m<sup>-3</sup>}}{\eqn{ug m^{-3}}}.
The risk categories match the Canadian AQHI and share the same health messaging:
\itemize{
\item Low:  AQHI+ of 1-3, or a concentration of 0-30 \ifelse{html}{\out{&mu;g m<sup>-3</sup>}}{\eqn{ug m^{-3}}}
\item Moderate: AQHI+ of 4-6, or a concentration of 30.1-60 \ifelse{html}{\out{&mu;g m<sup>-3</sup>}}{\eqn{ug m^{-3}}}
\item High: AQHI+ of 7-10, or a concentration of 60.1-100 \ifelse{html}{\out{&mu;g m<sup>-3</sup>}}{\eqn{ug m^{-3}}}
\item Very High: AQHI+ of +, or a concentration above 100 \ifelse{html}{\out{&mu;g m<sup>-3</sup>}}{\eqn{ug m^{-3}}}
}

The AQHI+ was originally published by Yao et al (2019): \doi{doi:10.17269/s41997-019-00237-w},
and has been adopted by all Canadian provinces/territories as of 2024.
(except Québec where they use the AQI instead of the AQHI/AQHI+).

Quote from Yao et al (2019):

\emph{This algorithm is a linear extrapolation based on two anchor points:
\enumerate{
\item the AQHI-Plus value should reach 7 (high risk category),
when 1-h PM2.5 concentrations are over 60 μg/m3.
This is consistent with [...] the analyses of the likelihood of exceeding the 25 μg/m3 24-h threshold value; and
\item the AQHI-Plus value should reach 4 (moderate risk category)
when the 1-h PM2.5 concentrations are over 30 μg/m3,
which is the threshold for odour perception of burning softwood
(Kistler et al. 2012: \doi{doi:10.1016/j.atmosenv.2012.01.044}),
the major type of forest species in BC.
}

[...]

These conclusions suggest that the AQHI is most indicative of the mortality and
circulatory risks during periods affected by wildfire smoke.
However, the [...] AQHI-Plus is most indicative of the respiratory risks
during high-intensity fire periods, particularly for people with asthma.
This is consistent with the rationale behind the AQHI formulation
and with existing health evidence related to the effects of wildfire smoke.}
}
\examples{
# Hourly pm2.5 concentrations
pm25 <- sample(1:150, 24)
# Calculate the detailed AQHI+ (tibble with risk levels etc)
AQHI_plus(pm25)
# Or just the AQHI+ (returned as a factor)
AQHI_plus(pm25, detailed = FALSE)

# Hourly pm2.5 concentrations (with negative values)
pm25 <- c(-2, -0.1, sample(1:150, 22))
# Calculate the AQHI+ for each hour, except for hours where pm2.5 is < -0.5
AQHI_plus(pm25, min_allowed_pm25 = -0.5)

}
\references{
Yao et al (2019): \doi{doi:10.17269/s41997-019-00237-w}

Environment and Climate Change Canada: \url{https://www.canada.ca/en/environment-climate-change/services/air-quality-health-index/about.html}
}
