% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AQHI_risk_categories}
\alias{AQHI_risk_categories}
\title{AQHI_risk_categories}
\format{
A named list of integer (or character for "Very High") vectors
}
\source{
\href{https://www.canada.ca/en/environment-climate-change/services/air-quality-health-index/about.html}{Environment and Climate ChangeD Canada}.
}
\usage{
AQHI_risk_categories
}
\description{
A named list of the risk categories and their respective AQHI levels.
}
\keyword{datasets}
