% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aqhi_colours.R
\name{get_aqhi_colours}
\alias{get_aqhi_colours}
\title{Get AQHI colours for each AQHI level}
\usage{
get_aqhi_colours(aqhi_levels = c(1:10, "+", NA))
}
\arguments{
\item{aqhi_levels}{(Optional).
A character or factor vector of AQHI levels (1-10, "+", or NA).
Default is all AQHI levels (1-10, "+", and NA).}
}
\value{
A character vector of hexidecimal codes correspoding to each AQHI level.
}
\description{
Get AQHI colours for each AQHI level
}
\examples{
get_aqhi_colours()

1:100 |> AQHI_plus(detailed = FALSE) |> get_aqhi_colours()
}
