% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_risk_category.R
\name{get_risk_category}
\alias{get_risk_category}
\title{Get the risk category for each AQHI level}
\usage{
get_risk_category(aqhi_levels, language = "en")
}
\arguments{
\item{aqhi_levels}{A vector of AQHI levels (1-10 or "+").}

\item{language}{(Optional).
A single character value indicating the language to use for risk levels and health messaging.
Must be either "en" (English) or "fr" (French). Not case sensitive.
Default is "en".}
}
\value{
A factor of risk categories (Low, Moderate, High, Very High) for each AQHI level.
}
\description{
Get the risk category for each AQHI level
}
\examples{
get_risk_category(1:10)
get_risk_category(1:10, language = "fr")
}
