% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQHI.R
\name{AQHI}
\alias{AQHI}
\title{Calculate the Canadian AQHI from hourly \ifelse{html}{\out{PM<sub>2.5</sub>}}{\eqn{PM_{2.5}}}, \ifelse{html}{\out{O<sub>3</sub>}}{\eqn{O_3}}, and \ifelse{html}{\out{NO<sub>2</sub>}}{\eqn{NO_2}} concentrations}
\usage{
AQHI(
  dates,
  pm25_1hr_ugm3,
  o3_1hr_ppb = NA_real_,
  no2_1hr_ppb = NA_real_,
  allow_aqhi_plus_override = TRUE,
  detailed = TRUE,
  language = "en",
  quiet = FALSE
)
}
\arguments{
\item{dates}{A POSIXct (date-time) vector of corresponding to hourly observations.
Date gaps will be filled with NA's automatically as needed when calculating 3-hour averages.}

\item{pm25_1hr_ugm3}{A numeric vector of hourly mean \ifelse{html}{\out{PM<sub>2.5</sub>}}{\eqn{PM_{2.5}}} concentrations (units = \ifelse{html}{\out{&mu;g m<sup>-3</sup>}}{\eqn{ug m^{-3}}}).}

\item{o3_1hr_ppb}{(Optional).
A numeric vector of hourly mean ozone (\ifelse{html}{\out{O<sub>3</sub>}}{\eqn{O_3}}) concentrations (units = ppb).
If all are NA (the default) AQHI won't be calculated and instead AQHI+ (see \code{\link{AQHI_plus}}) will be used.}

\item{no2_1hr_ppb}{(Optional).
A numeric vector of hourly mean nitrogen dioxide (\ifelse{html}{\out{NO<sub>2</sub>}}{\eqn{NO_2}}) concentrations (units = ppb).
If all are NA (the default) AQHI won't be calculated and instead AQHI+ (see \code{\link{AQHI_plus}}) will used.}

\item{allow_aqhi_plus_override}{(Optional).
A single logical value indicating if the AQHI+ should be allowed to override the AQHI if it exceeds the AQHI for a particular hour.
Default is TRUE.}

\item{detailed}{(Optional).
A single logical value indicating if a tibble with levels, risk categories, health messages, etc should be returned.
If FALSE only the levels will be returned.
Default is TRUE.}

\item{language}{(Optional).
A single character value indicating the language to use for risk levels and health messaging.
Must be either "en" (English) or "fr" (French). Not case sensitive.
Default is "en".}

\item{quiet}{(Optional).
A single logical (TRUE/FALSE) value indicating if non-critical warnings/messages should be hidden.
Default is FALSE.}
}
\value{
If \code{detailed = TRUE}:
\itemize{
\item A tibble (data.frame) with a detailed summary of the AQHI, risk levels, health messages, etc and the same number of rows as \code{length(dates)}
}

If \code{detailed = FALSE}:
\itemize{
\item a factor vector of AQHI levels with \code{length(dates)} elements
}
}
\description{
The Canadian Air Quality Health Index (AQHI) is a measure of the health risk associated with exposure to ambient air pollution.
It combines the impacts of fine particulate matter (\ifelse{html}{\out{PM<sub>2.5</sub>}}{\eqn{PM_{2.5}}}),
ozone (\ifelse{html}{\out{O<sub>3</sub>}}{\eqn{O_3}}),
and nitrogen dioxide (\ifelse{html}{\out{NO<sub>2</sub>}}{\eqn{NO_2}}) on a scale from 1-10 (+).

AQHI formula:
\if{html}{\out{<sup>}}10\if{html}{\out{</sup>}}⁄\if{html}{\out{<sub>}}10.4\if{html}{\out{</sub>}} × 100 ×
(e\if{html}{\out{<sup>}}0.000537 * O\if{html}{\out{<sub>}}3\if{html}{\out{</sub>}}\if{html}{\out{</sup>}} - 1) +
(e\if{html}{\out{<sup>}}0.000871 * NO\if{html}{\out{<sub>}}2\if{html}{\out{</sub>}}\if{html}{\out{</sup>}} - 1) +
(e\if{html}{\out{<sup>}}0.000487 * PM\if{html}{\out{<sub>}}2.5\if{html}{\out{</sub>}}\if{html}{\out{</sup>}} - 1)

The AQHI was originally published by Stieb et al. (2008): \doi{doi:10.3155/1047-3289.58.3.435},
and is used by all Canadian provinces and territories except for Québec, which uses the AQI instead.
The AQHI is calculated for each hour using 3-hour rolling mean concentrations of \ifelse{html}{\out{PM<sub>2.5</sub>}}{\eqn{PM_{2.5}}}, \ifelse{html}{\out{O<sub>3</sub>}}{\eqn{O_3}}, and \ifelse{html}{\out{NO<sub>2</sub>}}{\eqn{NO_2}}.
However, the AQHI is overriden by the AQHI+ for a given hour if that AQHI+ level is higher.
The AQHI+ is a modification of the AQHI that only uses \ifelse{html}{\out{PM<sub>2.5</sub>}}{\eqn{PM_{2.5}}} and is calculated using hourly means (see \code{\link{AQHI_plus}}).
}
\examples{
obs <- data.frame(
  date = seq(
    as.POSIXct("2024-01-01 00:00:00"),
    as.POSIXct("2024-01-01 23:00:00"), "1 hours"
  ),
  pm25 = sample(1:150, 24), o3 = sample(1:150, 24), no2 = sample(1:150, 24)
)

# Calculate the AQHI
obs$date |>
  AQHI(
    pm25_1hr_ugm3 = obs$pm25,
    o3_1hr_ppb = obs$o3,
    no2_1hr_ppb = obs$no2
  )

# Return just the AQHI levels
obs$date |>
  AQHI(
    pm25_1hr_ugm3 = obs$pm25,
    o3_1hr_ppb = obs$o3,
    no2_1hr_ppb = obs$no2,
    detailed = FALSE
  )

# Calculate just the AQHI+
obs$date |> AQHI(pm25_1hr_ugm3 = obs$pm25)
obs$date |> AQHI(pm25_1hr_ugm3 = obs$pm25, quiet = TRUE) # silence warning

# Return French version
obs$date |>
  AQHI(
    pm25_1hr_ugm3 = obs$pm25,
    o3_1hr_ppb = obs$o3,
    no2_1hr_ppb = obs$no2,
    language = "fr"
  )
obs$date |> AQHI(pm25_1hr_ugm3 = obs$pm25, language = "fr")

# Don't allow AQHI+ override (not recommended)
obs$date |>
  AQHI(
    pm25_1hr_ugm3 = obs$pm25,
    o3_1hr_ppb = obs$o3,
    no2_1hr_ppb = obs$no2,
    allow_aqhi_plus_override = FALSE
  )
}
\references{
Stieb et al. (2008): \doi{doi:10.3155/1047-3289.58.3.435}

Environment and Climate Change Canada: \url{https://www.canada.ca/en/environment-climate-change/services/air-quality-health-index/about.html}
}
