% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aqhi_colours.R
\name{get_aqhi_colours}
\alias{get_aqhi_colours}
\title{Get AQHI colours for each AQHI level}
\usage{
get_aqhi_colours(values = c(1:10, "+", NA), types = "aqhi")
}
\arguments{
\item{values}{(Optional).
A vector of AQHI levels (1-10, "+", or NA) if \code{type} is "aqhi" OR
A vector of hourly PM2.5 concentrations (ug/m3) if \code{type} is "pm25_1hr".
Default is all AQHI levels.}

\item{types}{(Optional).
A single character value, or a vector the same length as \code{values}, indicating the type(s) of values in \code{values}.
Must all be within either "aqhi" or "pm25_1hr".
Default is "aqhi".}
}
\value{
A character vector of hexidecimal codes correspoding to each level of \code{aqhi_levels}.
}
\description{
Every AQHI level (1-10, "+", and NA) is assigned a hexidecimal colour code for visualization.

These colours are defined by \href{https://www.canada.ca/en/environment-climate-change/services/air-quality-health-index/about.html}{Environment and Climate Change Canada} such that:
\itemize{
\item Low AQHI (1-3) are light to dark blue
\item Moderate AQHI (4-6) are yellow to orange
\item High AQHI (7-10) are light to dark red
\item Very High AQHI (+) is darker red
\item Missing AQHI (NA) is light grey
}
}
\examples{
# Get AQHI colours for all AQHI levels
get_aqhi_colours()

# Get AQHI colours for observation data
hourly_pm25_ugm3 <- sample(1:100, 50, replace = TRUE)
aqhi_levels <- hourly_pm25_ugm3 |>
  AQHI_plus(detailed = FALSE)
aqhi_levels |> get_aqhi_colours()

# The same but with PM2.5 provided
hourly_pm25_ugm3 |> get_aqhi_colours(types = "pm25_1hr")

# Or even a mix
values <- c(aqhi_levels, hourly_pm25_ugm3)
types <-  rep("aqhi", length(aqhi_levels)) |>
  c(rep("pm25_1hr", length(hourly_pm25_ugm3)))

values |> get_aqhi_colours(types = types)

}
\references{
Environment and Climate Change Canada: \url{https://www.canada.ca/en/environment-climate-change/services/air-quality-health-index/about.html}
}
