% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_health_messages.R
\name{get_health_messages}
\alias{get_health_messages}
\title{Get health messages for AQHI+ risk categories}
\usage{
get_health_messages(risk_categories, language = "en")
}
\arguments{
\item{risk_categories}{A factor or character vector of AQHI+ risk categories (Low, Moderate, High, Very High).}

\item{language}{(Optional).
A single character value indicating the language to use for risk levels and health messaging.
Must be either "en" (English) or "fr" (French). Not case sensitive.
Default is "en".}
}
\value{
A data frame of health messages (for at risk population and general population) for the AQHI+ risk categories.
}
\description{
Get health messages for AQHI+ risk categories
}
\examples{
get_health_messages(c("Low", "Moderate", "High", "Very High"))
get_health_messages(c("Low", "Moderate", "High", "Very High"), language = "fr")
}
