% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_risk_category.R
\name{get_risk_category}
\alias{get_risk_category}
\title{Get the risk category for each AQHI level}
\usage{
get_risk_category(aqhi_levels, language = "en")
}
\arguments{
\item{aqhi_levels}{A vector of AQHI levels (1-10 or "+").}

\item{language}{(Optional).
A single character value indicating the language to use for risk levels and health messaging.
Must be either "en" (English) or "fr" (French). Not case sensitive.
Default is "en".}
}
\value{
A factor of risk categories (Low, Moderate, High, Very High) representing each AQHI level.
}
\description{
AQHI levels are broken down into categories for easier communication of health risks:
\itemize{
\item Low: AQHI = 1-3
\item Moderate: AQHI = 4-6
\item High: AQHI = 7-10
\item Very High: AQHI = "+" (>10)
}

Health messages for at risk population and general population are available for each risk category - see \code{\link{get_health_messages}}
}
\examples{
# Get risk categories for all AQHI levels
get_risk_category()

# The same, but en Francais
get_risk_category(language = "fr")

# Get risk categories for obervations
aqhi_levels <- sample(c(1:10, "+"), 50, replace = TRUE)
get_risk_category(aqhi_levels)
}
\references{
Environment and Climate Change Canada: \url{https://www.canada.ca/en/environment-climate-change/services/air-quality-health-index/about.html}
}
